+++ core/src/global/ionic-global.ts
@@ -306,7 +306,7 @@ export const initialize = (userConfig: IonicConfig = {}) => {
          */
         if (shouldUseCloseWatcher()) {
           printIonWarning(
-            'experimentalCloseWatcher was set to `true`, but hardwareBackButton was set to `false`. Both config options must be `true` for the Close Watcher API to be used.'
+            '[ion-app] - experimentalCloseWatcher was set to `true`, but hardwareBackButton was set to `false`. Both config options must be `true` for the Close Watcher API to be used.'
           );
         }